Support updated login two factor auth flow
Flocknote introduced a new login flow that (1) grabs metadata about
the account based on the subdomain (getting the networkID), (2) asks
for the email address of the user logging in, (3) sends a code to the
email address with a link to optionally use a standard password instead.

These changes support this new flow.

Unfortunately, I cannot find a way to support this flow and not still
have a code sent via email.  I've tried a bunch of parameter changes
trying to turn this off, but without any luck.