Introduce Flockbot::Connection
This service class is designed to be a low-level interface around
Faraday HTTP client to make raw requests to Flocknote.

It also has a basic understanding of the Flocknote cookies that are
necessary for future requests.

This is the first step in unwinding some of the Flockbot classes so
that we can allow for reusing prior session cookies and support for
logging in using one time Flocknote codes.