Update commentary on unwanted one time passcodes
Unfortunately, I have been unable to avoid having one time passcodes
being sent during logging in with a password (where we are presumably
not even going to be using the one time passcode) and when logging in
with a one time passcode (where we don't actually want the second one).

I have tried all kinds of permeations of session setting and params
sending to avoid this, but have been unsuccessful.

My suspicion is that they are using a session token to track the
progression of the login form on the server, and so they know when
they have already sent a token or not.